+++ refinedstorage2-network-api/src/main/java/com/refinedmods/refinedstorage2/api/network/node/controller/ControllerListener.java
@@ -1,5 +1,0 @@
-package com.refinedmods.refinedstorage2.api.network.node.controller;
-
-public interface ControllerListener {
-    void onEnergyChanged();
-}