+++ refinedstorage2-platform-fabric/src/main/java/com/refinedmods/refinedstorage2/platform/fabric/integration/energy/DualEnergyStorage.java
@@ -1,0 +1,6 @@
+package com.refinedmods.refinedstorage2.platform.fabric.integration.energy;
+
+import com.refinedmods.refinedstorage2.api.network.energy.EnergyStorage;
+
+public interface DualEnergyStorage extends EnergyStorage, team.reborn.energy.api.EnergyStorage {
+}