+++ CHANGELOG.md
@@ -35,6 +35,7 @@ before updating.
   wrench by checking the `fabric:wrenches` tag.
 - Made energy usage by the Grid and Disk Drive less power-hungry.
 - The Controller now displays an energy bar on the item.
+- Upgrade Team Reborn Energy API.
 
 ### Removed